move graphql to lib. setguide after render. remove classes
